move new optim methods from wrong place
