+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/optim/AdamSpec.scala
@@ -1,12 +1,11 @@
 /*
- * Licensed to Intel Corporation under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * Intel Corporation licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
+ * Copyright 2016 The BigDL Authors.
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,20 +16,19 @@
 
 package com.intel.analytics.bigdl.optim
 
-import com.intel.analytics.bigdl.utils.T
-import org.scalatest.{FlatSpec, Matchers}
 import com.intel.analytics.bigdl.tensor.Tensor
+import com.intel.analytics.bigdl.utils.{T, TestUtils}
+import org.scalatest.{FlatSpec, Matchers}
 
 import scala.collection.mutable.ArrayBuffer
 
-// @com.intel.analytics.bigdl.tags.Parallel
-@com.intel.analytics.bigdl.tags.Serial
-class RMSpropSpec extends FlatSpec with Matchers {
+@com.intel.analytics.bigdl.tags.Parallel
+class AdamSpec extends FlatSpec with Matchers {
   val start = System.currentTimeMillis()
-  "RMSprop" should "perform well on rosenbrock function" in {
+  "adam" should "perform well on rosenbrock function" in {
     val x = Tensor[Double](2).fill(0)
-    val config = T("learningRate" -> 5e-4)
-    val optm = new RMSprop[Double]
+    val config = T("learningRate" -> 0.002)
+    val optm = new Adam[Double]
     var fx = new ArrayBuffer[Double]
     for (i <- 1 to 10001) {
       val result = optm.optimize(TestUtils.rosenBrock, x, config)
@@ -48,8 +46,9 @@ class RMSpropSpec extends FlatSpec with Matchers {
     val spend = System.currentTimeMillis() - start
     println("Time Cost: " + spend + "ms")
 
-    (fx.last < 1e-4) should be(true)
+    (fx.last < 1e-9) should be(true)
     x(Array(1)) should be(1.0 +- 0.01)
     x(Array(2)) should be(1.0 +- 0.01)
   }
 }
+
